moved dashboard to pages folder
